requirements.txt
earthengine-api==0.1.331
folium==0.13.0
geemap==0.17.2
streamlit==1.14.1
streamlit_ext==0.1.4
streamlit-folium==0.7.0
pre-commit==2.18.1
altair<5
python-dotenv==1.0.1